magmax vs galah test

+++ src/assess.rs
@@ -9,8 +9,8 @@ use log::error;
 #[derive(Clone)]
 #[derive(Debug)]
 pub struct BinQuality {
-    pub completeness: f64,
-    pub contamination: f64,
+    pub completeness: f32,
+    pub contamination: f32,
 }
 
 /// Run CheckM2 to obtain completeness and contamination of input bins
@@ -84,8 +84,8 @@ pub fn parse_bins_quality(
                 break;
             }
         }
-        let completeness: f64 = record[1].parse().unwrap_or(0.0);
-        let contamination: f64 = record[2].parse().unwrap_or(0.0);
+        let completeness: f32 = record[1].parse().unwrap_or(0.0);
+        let contamination: f32 = record[2].parse().unwrap_or(0.0);
         bin_qualities.insert(bin_id, 
         BinQuality{ completeness, contamination });
     }
